requirements.txt
flask==3.1.1
mysql-connector-python==9.3.0
python-dotenv==1.1.0
